+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Remove error logs
+
 ### 1.6.0: June 17th, 2021
 * Remove special chars from tokens
 * Aquos contact feedback with name and booking date