+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Remove error logs
+* Fix bookings query with prepare, that prevented case insensitive searches
 
 ### 1.6.0: June 17th, 2021
 * Remove special chars from tokens